Renaming bowline package to exposedcmd
